+++ cypress/e2e/ps1784/03_mollie.ps1784.Subscriptions.WIP.js
@@ -17,7 +17,7 @@ const login = (MollieBOFOLoggingIn) => {
   })
   }
 
-//Checing the console for errors
+//Checking the console for errors
 let windowConsoleError;
 Cypress.on('window:before:load', (win) => {
 windowConsoleError = cy.spy(win.console, 'error');
@@ -30,31 +30,32 @@ describe('PS1784 Subscriptions Test Suit', () => {
       cy.viewport(1920,1080)
       login('MollieBOFOLoggingIn')
   })
-it.skip('C176305 Check if Subscription options added in Product BO', () => {
+it('C176305 Check if Subscription options added in Product BO', () => {
   cy.visit('/admin1/')
   cy.get('#subtab-AdminCatalog > :nth-child(1)').click()
   cy.get('#subtab-AdminProducts > .link').click()
-  cy.contains('Hummingbird printed t-shirt').click()
-  cy.get('[name="show_variations"]').check()
+  cy.get('[data-product-id="8"]').find('[class="btn tooltip-link product-edit"]').click()
+  cy.contains('Product with combinations').click()
   cy.get('[id="tab_step3"]').click()
   cy.contains('Daily').click({force:true})
   cy.get('[class="token"]').should('be.visible')
   cy.get('#create-combinations').click()
   cy.wait(5000)
-  cy.contains('Mollie subscription - Daily').should('be.visible')
+  cy.reload()
   cy.wait(5000)
+  cy.contains('Mollie Subscription - Daily').should('be.visible')
   cy.get('[class="attribute-quantity"]').last().find('[type="text"]').clear().type('999')
   cy.get('#submit').click()
   cy.get('.growl-message').contains('Settings updated.')
   //Check if Subscription options are in Product Page FO
   cy.visit('/SHOP2/de/')
-  cy.get('.products > :nth-child(1)').click()
+  cy.get('[data-id-product="8"]').click() //possible PS1784 notice exception, checking with dev...
   cy.get('a').click()
   //wip ...
   //Check if Subscription options are implemented in My Account FO
   cy.visit('/SHOP2/')
   cy.get('[class="account"]').click()
-  cy.contains('Mollie subscriptions').click()
+  cy.contains('Subscriptions').click()
   cy.get('[class="page-content"]').should('be.visible')
   //wip ...
 });