+++ src/Application/CommandHandler/UpdateApplePayShippingMethodHandler.php
@@ -15,35 +15,41 @@ namespace Mollie\Application\CommandHandler;
 use Cart;
 use Mollie\Application\Command\UpdateApplePayShippingMethod;
 use Mollie\Config\Config;
-use Mollie\Service\OrderFeeService;
+use Mollie\Service\OrderPaymentFeeService;
 
 final class UpdateApplePayShippingMethodHandler
 {
     /**
-     * @var OrderFeeService
+     * @var OrderPaymentFeeService
      */
-    private $orderFeeService;
+    private $orderPaymentFeeService;
 
-    public function __construct(OrderFeeService $orderFeeService)
+    public function __construct(OrderPaymentFeeService $orderPaymentFeeService)
     {
-        $this->orderFeeService = $orderFeeService;
+        $this->orderPaymentFeeService = $orderPaymentFeeService;
     }
 
     public function handle(UpdateApplePayShippingMethod $command): array
     {
         $cart = new Cart($command->getCartId());
+
         $cart->id_carrier = $command->getCarrierId();
         $cart->setDeliveryOption([
            $cart->id_address_delivery => $command->getCarrierId() . ',',
        ]);
+
         $cart->update();
-        $orderTotal = $cart->getOrderTotal(true, Cart::BOTH, null, $command->getCarrierId());
-        $fee = $this->orderFeeService->getPaymentFee($orderTotal, Config::APPLEPAY);
+
+        $orderTotal = (float) $cart->getOrderTotal(true, Cart::BOTH, null, $command->getCarrierId());
+
+        $paymentFeeData = $this->orderPaymentFeeService->getPaymentFee($orderTotal, Config::APPLEPAY);
+
+        $paymentFee = $paymentFeeData->getPaymentFeeTaxIncl();
 
         return [
            'success' => true,
-           'data' => [
-               'amount' => $orderTotal + $fee,
+           'data' => [// TODO use calculator
+               'amount' => $orderTotal + $paymentFee,
            ],
        ];
     }