+++ src/Provider/PaymentOption/IdealPaymentOptionProvider.php
@@ -40,6 +40,7 @@ use Mollie;
 use Mollie\Adapter\LegacyContext;
 use Mollie\Builder\Content\PaymentOption\IdealDropdownInfoBlock;
 use Mollie\Provider\CreditCardLogoProvider;
+use Mollie\Provider\OrderTotal\OrderTotalProviderInterface;
 use Mollie\Provider\PaymentFeeProviderInterface;
 use Mollie\Service\Content\TemplateParserInterface;
 use Mollie\Service\LanguageService;
@@ -85,6 +86,8 @@ class IdealPaymentOptionProvider implements PaymentOptionProviderInterface
      * @var LanguageService
      */
     private $languageService;
+    /** @var OrderTotalProviderInterface */
+    private $orderTotalProvider;
 
     public function __construct(
         Mollie $module,
@@ -93,7 +96,8 @@ class IdealPaymentOptionProvider implements PaymentOptionProviderInterface
         PaymentFeeProviderInterface $paymentFeeProvider,
         TemplateParserInterface $templateParser,
         IdealDropdownInfoBlock $idealDropdownInfoBlock,
-        LanguageService $languageService
+        LanguageService $languageService,
+        OrderTotalProviderInterface $orderTotalProvider
     ) {
         $this->module = $module;
         $this->context = $context;
@@ -102,14 +106,16 @@ class IdealPaymentOptionProvider implements PaymentOptionProviderInterface
         $this->templateParser = $templateParser;
         $this->idealDropdownInfoBlock = $idealDropdownInfoBlock;
         $this->languageService = $languageService;
+        $this->orderTotalProvider = $orderTotalProvider;
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getPaymentOption(MolPaymentMethod $paymentMethod)
+    public function getPaymentOption(MolPaymentMethod $paymentMethod): PaymentOption
     {
         $paymentOption = new PaymentOption();
+
         $paymentOption->setCallToActionText(
             $paymentMethod->title ?:
             $this->languageService->lang($paymentMethod->method_name)
@@ -136,9 +142,9 @@ class IdealPaymentOptionProvider implements PaymentOptionProviderInterface
         ));
         $paymentOption->setLogo($this->creditCardLogoProvider->getMethodOptionLogo($paymentMethod));
 
-        $paymentFee = $this->paymentFeeProvider->getPaymentFee($paymentMethod);
+        $paymentFeeData = $this->paymentFeeProvider->getPaymentFee($paymentMethod, $this->orderTotalProvider->getOrderTotal());
 
-        if ($paymentFee) {
+        if ($paymentFeeData->isActive()) {
             $paymentOption->setInputs(
                 [
                     [
@@ -149,12 +155,20 @@ class IdealPaymentOptionProvider implements PaymentOptionProviderInterface
                     [
                         'type' => 'hidden',
                         'name' => 'payment-fee-price',
-                        'value' => $paymentFee,
+                        'value' => $paymentFeeData->getPaymentFeeTaxIncl(),
                     ],
                     [
                         'type' => 'hidden',
                         'name' => 'payment-fee-price-display',
-                        'value' => sprintf($this->module->l('Payment Fee: %1s', self::FILE_NAME), Tools::displayPrice($paymentFee)),
+                        'value' => sprintf(
+                            $this->module->l('Payment Fee: %1s', self::FILE_NAME),
+                            Tools::displayPrice($paymentFeeData->getPaymentFeeTaxIncl())
+                        ),
+                    ],
+                    [
+                        'type' => 'hidden',
+                        'name' => 'payment-method-id',
+                        'value' => $paymentMethod->id,
                     ],
                 ]
             );