+++ src/Service/CartLinesService.php
@@ -13,10 +13,12 @@
 namespace Mollie\Service;
 
 use Cart;
+use Mollie\Adapter\Context;
 use Mollie\Adapter\ToolsAdapter;
 use Mollie\Config\Config;
 use Mollie\DTO\Line;
 use Mollie\DTO\Object\Amount;
+use Mollie\DTO\PaymentFeeData;
 use Mollie\Utility\CalculationUtility;
 use Mollie\Utility\CartPriceUtility;
 use Mollie\Utility\NumberUtility;
@@ -38,17 +40,19 @@ class CartLinesService
      * @var ToolsAdapter
      */
     private $tools;
+    private $context;
 
-    public function __construct(LanguageService $languageService, VoucherService $voucherService, ToolsAdapter $tools)
+    public function __construct(LanguageService $languageService, VoucherService $voucherService, ToolsAdapter $tools, Context $context)
     {
         $this->voucherService = $voucherService;
         $this->languageService = $languageService;
         $this->tools = $tools;
+        $this->context = $context;
     }
 
     /**
      * @param float $amount
-     * @param float $paymentFee
+     * @param PaymentFeeData $paymentFeeData
      * @param string $currencyIsoCode
      * @param array $cartSummary
      * @param float $shippingCost
@@ -62,7 +66,7 @@ class CartLinesService
      */
     public function getCartLines(
         $amount,
-        $paymentFee,
+        $paymentFeeData,
         $currencyIsoCode,
         $cartSummary,
         $shippingCost,
@@ -70,6 +74,8 @@ class CartLinesService
         $psGiftWrapping,
         $selectedVoucherCategory
     ) {
+        // TODO refactor whole service, split order line append into separate services and test them individually at least!!!
+
         $apiRoundingPrecision = Config::API_ROUNDING_PRECISION;
         $vatRatePrecision = Config::VAT_RATE_ROUNDING_PRECISION;
 
@@ -108,7 +114,7 @@ class CartLinesService
         $orderLines = $this->addWrappingLine($wrappingPrice, $cartSummary, $vatRatePrecision, $apiRoundingPrecision, $orderLines);
 
         // Add fee
-        $orderLines = $this->addPaymentFeeLine($paymentFee, $apiRoundingPrecision, $orderLines);
+        $orderLines = $this->addPaymentFeeLine($paymentFeeData, $apiRoundingPrecision, $orderLines);
 
         // Ungroup all the cart lines, just one level
         $newItems = $this->ungroupLines($orderLines);
@@ -194,6 +200,8 @@ class CartLinesService
                         'unitPrice' => 0,
                         'totalAmount' => 0,
                         'category' => '',
+                        'product_url' => $this->context->getProductLink($cartItem['id_product']),
+                        'image_url' => $this->context->getImageLink($cartItem['link_rewrite'], $cartItem['id_image']),
                     ];
                     continue;
                 }
@@ -212,6 +220,8 @@ class CartLinesService
                 'unitPrice' => round($cartItem['price_wt'], $apiRoundingPrecision),
                 'totalAmount' => (float) $roundedTotalWithTax,
                 'category' => $this->voucherService->getVoucherCategory($cartItem, $selectedVoucherCategory),
+                'product_url' => $this->context->getProductLink($cartItem['id_product']),
+                'image_url' => $this->context->getImageLink($cartItem['link_rewrite'], $cartItem['id_image']),
             ];
             $remaining -= $roundedTotalWithTax;
         }
@@ -334,6 +344,8 @@ class CartLinesService
                     'totalAmount' => round($totalAmount, $apiRoundingPrecision),
                     'vatRate' => round($actualVatRate, $apiRoundingPrecision),
                     'vatAmount' => round($vatAmount, $apiRoundingPrecision),
+                    'product_url' => $line['product_url'] ?? null,
+                    'image_url' => $line['image_url'] ?? null,
                 ];
                 if (isset($line['sku'])) {
                     $newItem['sku'] = $line['sku'];
@@ -410,27 +422,29 @@ class CartLinesService
     }
 
     /**
-     * @param float $paymentFee
+     * @param PaymentFeeData $paymentFeeData
      * @param int $apiRoundingPrecision
      * @param array $orderLines
      *
      * @return array
      */
-    private function addPaymentFeeLine($paymentFee, $apiRoundingPrecision, array $orderLines)
+    private function addPaymentFeeLine($paymentFeeData, $apiRoundingPrecision, array $orderLines)
     {
-        if ($paymentFee) {
-            $orderLines['surcharge'] = [
-                [
-                    'name' => $this->languageService->lang('Payment fee'),
-                    'sku' => Config::PAYMENT_FEE_SKU,
-                    'quantity' => 1,
-                    'unitPrice' => round($paymentFee, $apiRoundingPrecision),
-                    'totalAmount' => round($paymentFee, $apiRoundingPrecision),
-                    'vatAmount' => 0,
-                    'vatRate' => 0,
-                ],
-            ];
+        if (!$paymentFeeData->isActive()) {
+            return $orderLines;
         }
+
+        $orderLines['surcharge'] = [
+            [
+                'name' => $this->languageService->lang('Payment fee'),
+                'sku' => Config::PAYMENT_FEE_SKU,
+                'quantity' => 1,
+                'unitPrice' => round($paymentFeeData->getPaymentFeeTaxIncl(), $apiRoundingPrecision),
+                'totalAmount' => round($paymentFeeData->getPaymentFeeTaxIncl(), $apiRoundingPrecision),
+                'vatAmount' => NumberUtility::minus($paymentFeeData->getPaymentFeeTaxIncl(), $paymentFeeData->getPaymentFeeTaxExcl()),
+                'vatRate' => $paymentFeeData->getTaxRate(),
+            ],
+        ];
 
         return $orderLines;
     }
@@ -497,6 +511,8 @@ class CartLinesService
             }
 
             $line->setVatRate(TextFormatUtility::formatNumber($item['vatRate'], $apiRoundingPrecision, '.', ''));
+            $line->setProductUrl($item['product_url'] ?? null);
+            $line->setImageUrl($item['image_url'] ?? null);
 
             $newItems[$index] = $line;
         }